# docs/architecture/openapi.yaml
...
openapi: 3.1.0
info:
  title: API Análisis Predictivo — Desarrollos Comerciales del Sur, S.A.
  description: Contrato del prototipo de tesis. Autenticación JWT y control de acceso RBAC.
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Desarrollo local
paths:
...
  /api/v1/products:
    get:
      tags:
      - Products
      summary: Listar productos
      description: 'Requiere `inventario:leer`. Por defecto solo activos (`activo=true`); `activo=false`
        lista las bajas. Filtros: `categoria_id` y `q` (SKU o nombre).'
      operationId: listar_productos_api_v1_products_get
      security:
      - HTTPBearer: []
      parameters:
      - name: activo
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          default: true
          title: Activo
      - name: categoria_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Categoria Id
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Products
      summary: Crear producto
      description: Requiere `productos:crear`. Crea también su existencia en 0. 409 `SKU_DUPLICADO`; 404
        `CATEGORIA_NO_ENCONTRADA` / `PROVEEDOR_NO_ENCONTRADO`.
      operationId: crear_producto_api_v1_products_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/products/{producto_id}:
    get:
      tags:
      - Products
      summary: Detalle de un producto
      description: Requiere `inventario:leer`. 404 `PRODUCTO_NO_ENCONTRADO`.
      operationId: obtener_producto_api_v1_products__producto_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: producto_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Producto Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Products
      summary: Actualizar producto (parcial)
      description: Requiere `productos:editar`. No modifica existencias ni estado. 404 `PRODUCTO_NO_ENCONTRADO`;
        409 `SKU_DUPLICADO`.
      operationId: actualizar_producto_api_v1_products__producto_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: producto_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Producto Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Products
      summary: Dar de baja un producto (baja lógica)
      description: Requiere `productos:eliminar`. Marca `activo=false`; el historial se conserva. Idempotente.
        409 `PRODUCTO_EN_USO` si tiene stock reservado o cargas vigentes.
      operationId: eliminar_producto_api_v1_products__producto_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: producto_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Producto Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/products/{producto_id}/reactivate:
    post:
      tags:
      - Products
      summary: Reactivar un producto dado de baja
      description: Requiere `productos:eliminar`. Idempotente. 404 `PRODUCTO_NO_ENCONTRADO`.
      operationId: reactivar_producto_api_v1_products__producto_id__reactivate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: producto_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Producto Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProductoCreate:
      properties:
        sku:
          type: string
          maxLength: 30
          minLength: 1
          title: Sku
        nombre:
          type: string
          maxLength: 200
          minLength: 2
          title: Nombre
        categoria_id:
          type: string
          format: uuid
          title: Categoria Id
        proveedor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Proveedor Id
        unidad_medida:
          type: string
          maxLength: 20
          minLength: 1
          title: Unidad Medida
          default: unidad
        precio_venta:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Precio Venta
          default: '0'
        costo_unitario:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          title: Costo Unitario
          default: '0'
        stock_minimo:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,10}|(?=[\d.]{1,13}0*$)\d{0,10}\.\d{0,2}0*$)
          title: Stock Minimo
          default: '0'
      type: object
      required:
      - sku
      - nombre
      - categoria_id
      title: ProductoCreate
    ProductoOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        sku:
          type: string
          title: Sku
        nombre:
          type: string
          title: Nombre
        categoria_id:
          type: string
          format: uuid
          title: Categoria Id
        categoria:
          type: string
          title: Categoria
        proveedor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Proveedor Id
        unidad_medida:
          type: string
          title: Unidad Medida
        precio_venta:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Precio Venta
        costo_unitario:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Costo Unitario
        stock_minimo:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Stock Minimo
        activo:
          type: boolean
          title: Activo
        stock_actual:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Stock Actual
        stock_reservado:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Stock Reservado
        stock_disponible:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Stock Disponible
        creado_en:
          type: string
          format: date-time
          title: Creado En
      type: object
      required:
      - id
      - sku
      - nombre
      - categoria_id
      - categoria
      - unidad_medida
      - precio_venta
      - costo_unitario
      - stock_minimo
      - activo
      - stock_actual
      - stock_reservado
      - stock_disponible
      - creado_en
      title: ProductoOut
    ProductoPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProductoOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: ProductoPage
    ProductoUpdate:
      properties:
        sku:
          anyOf:
          - type: string
            maxLength: 30
            minLength: 1
          - type: 'null'
          title: Sku
        nombre:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 2
          - type: 'null'
          title: Nombre
        categoria_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Categoria Id
        proveedor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Proveedor Id
        unidad_medida:
          anyOf:
          - type: string
            maxLength: 20
            minLength: 1
          - type: 'null'
          title: Unidad Medida
        precio_venta:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          - type: 'null'
          title: Precio Venta
        costo_unitario:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,12}|(?=[\d.]{1,15}0*$)\d{0,12}\.\d{0,2}0*$)
          - type: 'null'
          title: Costo Unitario
        stock_minimo:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,10}|(?=[\d.]{1,13}0*$)\d{0,10}\.\d{0,2}0*$)
          - type: 'null'
          title: Stock Minimo
      type: object
      title: ProductoUpdate
      description: 'Actualización parcial. No modifica existencias (usar `POST /inventory/adjustments`)
        ni el

        estado (usar `DELETE` / `POST .../reactivate`).'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer